feat: appendable
- add appendable
- add return error for each
