+++ hashmap/core.go
