+++ hashmap/core_benchmark_test.go
