+++ hashmap/core_test.go
